v1.3.0
feat(Store): Added 'schema' getter and setter.
feat(Node): Add node class reference to internal metadata for instantiating child nodes. Useful if a store wants to use an extended 'Node' class.
refactor(Store): Changed 'Store.load()' to only take a 'file'. Not counting this as a breaking change as it's an internal API.